reorganize unit tests
