--- catalog.json
@@ -1,76 +1,0 @@
-[{
-    "pk": 1,
-    "model": "catalog.source",
-    "fields": {
-        "platform": 1,
-        "instrument": 2,
-        "specs": "Nothing special"
-    }
-},{
-    "pk": 2,
-    "model": "catalog.source",
-    "fields": {
-        "platform": 2,
-        "instrument": 2,
-        "specs": "Nothing special"
-    }
-},{
-    "pk": 1,
-    "model": "catalog.geographiclocation",
-    "fields": {
-        "geometry": "SRID=4326;POLYGON ((0.0000000000000000 0.0000000000000000, 0.0000000000000000 10.0000000000000000, 10.0000000000000000 10.0000000000000000, 10.0000000000000000 0.0000000000000000, 0.0000000000000000 0.0000000000000000))"
-    }
-},{
-    "pk": 2,
-    "model": "catalog.geographiclocation",
-    "fields": {
-        "geometry": "SRID=4326;POLYGON ((20.0000000000000000 20.0000000000000000, 20.0000000000000000 30.0000000000000000, 30.0000000000000000 30.0000000000000000, 30.0000000000000000 20.0000000000000000, 20.0000000000000000 20.0000000000000000))"
-    }
-},{
-    "pk": 1,
-    "model": "catalog.dataset",
-    "fields": {
-        "entry_id": "NERSC_test_dataset_titusen",
-        "entry_title": "Test dataset",
-        "time_coverage_end": "2010-01-02T00:00:00Z",
-        "time_coverage_start": "2010-01-01T00:00:00Z",
-        "summary": "This is a quite short summary about the test dataset.",
-        "source": 1,
-        "geographic_location": 1,
-        "data_center": 1,
-        "gcmd_location": 1,
-        "ISO_topic_category": 1,
-        "access_constraints": null
-    }
-},{
-    "pk": 2,
-    "model": "catalog.dataset",
-    "fields": {
-        "entry_id": "NERSC_test_dataset_tjuetusen",
-        "entry_title": "Test child dataset",
-        "time_coverage_end": "2010-01-03T00:00:00Z",
-        "time_coverage_start": "2010-01-02T00:00:00Z",
-        "summary": "This is a quite short summary about the test dataset.",
-        "source": 2,
-        "geographic_location": 2,
-        "data_center": 2,
-        "gcmd_location": 1,
-        "ISO_topic_category": 1,
-        "access_constraints": null
-    }
-},{
-    "model": "catalog.dataseturi",
-    "pk": 1,
-    "fields": {
-        "uri": "file://localhost/some/test/file1.ext",
-        "dataset": 1
-    }
-},{
-    "model": "catalog.dataseturi",
-    "pk": 2,
-    "fields": {
-        "uri": "file://localhost/some/test/file2.ext",
-        "dataset": 2
-    }
-}
-]
