+++ geospaas_rest_api/tests/urls.py
@@ -1,0 +1,8 @@
+"""geospaas_rest_api testing URL Configuration"""
+
+from django.conf.urls import include, url
+
+urlpatterns = [
+    url(r'^', include('geospaas.urls')),
+    url(r'^api/', include('geospaas_rest_api.urls')),
+]
